+++ ktrain/text/qa/core.py
@@ -279,9 +279,15 @@ class SimpleQA(QA):
           list
         """
         # locate candidate document contexts
-        doc_results = self.search(question, limit=n_docs_considered)
         paragraphs = []
         refs = []
+        doc_results = self.search(question, limit=n_docs_considered)
+        # TODO: tokenize by default?
+        if not doc_results: 
+            doc_results = self.search(TU.tokenize(question, join_tokens=True), limit=n_docs_considered)
+            if not doc_results:
+                warnings.warn('No documents matched words in question')
+                return []
         for doc_result in doc_results:
             rawtext = doc_result.get('rawtext', '')
             reference = doc_result.get('reference', '')