+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Most recent releases are shown at the top. Each release shows:
 - N/A
 
 ### Fixed:
+- return gracefully if no documents match question in `qa` module
+- tokenize question in `qa` module to prevent zero results
 - Added error in `text.preprocessor` when training set has incomplete integer labels